Ajouté vérification de propriété du groupe de mots de passes.

+++ passwords/views.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from django.shortcuts import render_to_response, get_object_or_404
+from django.shortcuts import render_to_response, get_object_or_404, redirect
 from django.http import HttpResponseRedirect, HttpResponse
 from django.core.urlresolvers import reverse
 from django.template import RequestContext
@@ -22,6 +22,11 @@ def index(request):
 @login_required()
 def group(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, group=group_id).user:
+        return redirect('/passwords')
+
     password_list = Password.objects.all().filter(group=group).order_by('name')
     return render(request, 'passwords/group.html', {'group': group, 'password_list': password_list})
 
@@ -37,15 +42,25 @@ def create_group(request):
     group.save()
     account = Account(user=request.user, group=group).save()
     return index(request)
-    
+
 @login_required()
 def edit_group(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, group=group_id).user:
+        return redirect('/passwords')
+
     return render(request, 'passwords/edit_group.html', {'group': group}, context=True)
 
 @login_required()
 def save_group(request, group_id):
     grp = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, grp=group_id).user:
+        return redirect('/passwords')
+
     grp.name = request.POST['Name']
     grp.details = request.POST['Details']
     grp.save()
@@ -54,6 +69,11 @@ def save_group(request, group_id):
 @login_required()
 def do_delete_group(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, group=group_id).user:
+        return redirect('/passwords')
+
     passwords = Password.objects.all().filter(group=group).count()
     if passwords == 0:
         group.delete()
@@ -70,17 +90,32 @@ def do_delete_group(request, group_id):
 @login_required()
 def password(request, group_id, password_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, group=group_id).user:
+        return redirect('/passwords')
+
     password = get_object_or_404(Password, pk=password_id)
     return render(request, 'passwords/password.html', {'group': group, 'password': password})
 
 @login_required()
 def new_password(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, group=group_id).user:
+        return redirect('/passwords')
+
     return render(request, 'passwords/new_password.html', {'group': group}, context=True)
 
 @login_required()
 def create_password(request, group_id):
     grp = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, grp=group_id).user:
+        return redirect('/passwords')
+
     pwd = Password()
     pwd.name = request.POST['Name']
     pwd.details = request.POST['Details']
@@ -93,6 +128,11 @@ def create_password(request, group_id):
 @login_required()
 def edit_password(request, group_id, password_id):
     group = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, grp=group_id).user:
+        return redirect('/passwords')
+
     password = get_object_or_404(Password, pk=password_id)
     forum_list = Forum.objects.all().order_by('name')
     return render(request, 'passwords/edit_password.html', {'group': group, 'password': password}, context=True)
@@ -100,6 +140,11 @@ def edit_password(request, group_id, password_id):
 @login_required()
 def save_password(request, group_id, password_id):
     grp = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, grp=group_id).user:
+        return redirect('/passwords')
+
     pwd = get_object_or_404(Password, pk=password_id)
     pwd.name = request.POST['Name']
     pwd.details = request.POST['Details']
@@ -111,6 +156,11 @@ def save_password(request, group_id, password_id):
 @login_required()
 def do_delete_password(request, group_id, password_id):
     grp = get_object_or_404(Group, pk=group_id)
+
+    # Vérifie que l'utilisateur a accès à ce groupe
+    if request.user != get_object_or_404(Account, grp=group_id).user:
+        return redirect('/passwords')
+
     pwd = get_object_or_404(Password, pk=password_id)
     pwd.delete()
     return group(request, group_id)
